Move getMessageLabelStrings() to pmapi/messages.go
The function has been moved here and is called from
client::GetMessage(). The message is fetched and then the LabelNames
are assigned from the result of calling this function.

This prevents label names from being fetched during tests.
GetMessage() is mocked and therefore the inner call to this function
is not executed during tests.